test(app): add testAppMockNotificationService
 - mocks out the NotificationService so that tests don't get hung up on needing to click to dismiss any generated notifications.
